Diverse wijzigingen

+++ lib/event_handler/src/event_handler.cpp
@@ -8,7 +8,7 @@ namespace NsEvent {
     /*****************************************************/
 
  
-    void MouseHandler::initHandler(int typeMorseKey, int leftEvent, int rightEvent) {
+    void MouseHandler::initHandler(u_int8_t  typeMorseKey, u_int8_t  leftEvent, u_int8_t  rightEvent) {
 
             this->typeMorseKey = typeMorseKey;
             this->leftEvent    = leftEvent;
@@ -19,13 +19,14 @@ namespace NsEvent {
 
     void MouseHandler::handleMorseKey(int pinDot ,int pinDash) {
     
-             if(pinDot == LOW)                      { //pin==LOW  => is closed (activated )
-                 Mouse.press(this->leftEvent);              // print a space on the monitor window    
-              }else if(pinDash == LOW)              { //pin==LOW  => is closed (activated )  
-                 Mouse.press(this->rightEvent);             // print a space on the monitor window    
-              }  else {                               //pin==HIGH => is open ( deactived )
+             if(pinDot == LOW)                      {         //pin==LOW  => is closed (activated )
+                 Mouse.press(this->leftEvent);              
+              }else if(pinDash == LOW)              {         //pin==LOW  => is closed (activated )  
+                 Mouse.press(this->rightEvent);     
+              }  else {                                       //pin==HIGH => is open ( deactived )
                  Mouse.release();
-              }  
+              } 
+               
      }
 
 
@@ -35,7 +36,7 @@ namespace NsEvent {
     /*****************************************************/
 
   
-    void KeyboardHandler::initHandler(int typeMorseKey, int leftEvent, int rightEvent)  {
+    void KeyboardHandler::initHandler(u_int8_t  typeMorseKey, u_int8_t  leftEvent, u_int8_t  rightEvent)  {
 
             this->typeMorseKey = typeMorseKey;
             this->leftEvent    = leftEvent;
@@ -48,9 +49,9 @@ namespace NsEvent {
     void KeyboardHandler::handleMorseKey(int pinDot ,int pinDash) {
       
              if(pinDot == LOW)                      { //pin==LOW  => is closed (activated )
-                 Keyboard.press(this->leftEvent);           // print a space on the monitor window    
+                 Keyboard.press(this->leftEvent);    
               }else if(pinDash == LOW)              { //pin==LOW  => is closed (activated )  
-                 Keyboard.press(this->rightEvent);          // print a space on the monitor window    
+                 Keyboard.press(this->rightEvent);    
               }  else {                               //pin==HIGH => is open ( deactived )
                  Keyboard.releaseAll();
               }
